Update deps, split stuck items
