+++ .travis.yml
@@ -7,19 +7,19 @@ jobs:
     if: tag IS present
     osx_image: xcode13.2
     language: go
-    go: 1.17.x
+    go: 1.19.x
   - if: tag IS NOT present AND branch != unstable
     os: linux
     dist: bionic
     services: docker
     language: go
-    go: 1.17.x
+    go: 1.19.x
   - if: tag IS present OR branch = unstable
     os: linux
     dist: bionic
     services: docker
     language: go
-    go: 1.17.x
+    go: 1.19.x
     env: MAKEBUILD=release
 git:
   depth: false
