+++ pkg/triggers/stuckitems/lidarr.go
@@ -1,0 +1,52 @@
+package stuckitems
+
+import (
+	"strings"
+	"time"
+)
+
+func (c *cmd) getFinishedItemsLidarr() ItemList { //nolint:dupl,cyclop
+	stuck := make(ItemList)
+
+	for idx, app := range c.Apps.Lidarr {
+		instance := idx + 1
+
+		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
+			continue
+		}
+
+		if app.Lidarr == nil {
+			c.Errorf("Getting Lidarr Queue (%d): Lidarr config is nil? This is probably a bug.", instance)
+			continue
+		}
+
+		start := time.Now()
+
+		queue, err := app.GetQueue(getItemsMax, getItemsMax)
+		if err != nil {
+			c.Errorf("Getting Lidarr Queue (%d): %v", instance, err)
+			continue
+		}
+
+		stuckapp := stuck[instance]
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			}
+
+			item.Quality = nil
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = app.Name
+		stuckapp.Elapsed = time.Since(start)
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Lidarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
+}
