+++ pkg/triggers/stuckitems/radarr.go
@@ -1,0 +1,54 @@
+package stuckitems
+
+import (
+	"strings"
+	"time"
+)
+
+func (c *cmd) getFinishedItemsRadarr() ItemList { //nolint:cyclop
+	stuck := make(ItemList)
+
+	for idx, app := range c.Apps.Radarr {
+		instance := idx + 1
+
+		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
+			continue
+		}
+
+		if app.Radarr == nil {
+			c.Errorf("Getting Radarr Queue (%d): Radarr config is nil? This is probably a bug.", instance)
+			continue
+		}
+
+		start := time.Now()
+
+		queue, err := app.GetQueue(getItemsMax, 1)
+		if err != nil {
+			c.Errorf("Getting Radarr Queue (%d): %v", instance, err)
+			continue
+		}
+
+		stuckapp := stuck[instance]
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			}
+
+			item.Quality = nil
+			item.CustomFormats = nil
+			item.Languages = nil
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = app.Name
+		stuckapp.Elapsed = time.Since(start)
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Radarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
+}
