+++ pkg/triggers/stuckitems/sonarr.go
@@ -1,0 +1,60 @@
+package stuckitems
+
+import (
+	"strings"
+	"time"
+
+	"golift.io/starr/sonarr"
+)
+
+func (c *cmd) getFinishedItemsSonarr() ItemList { //nolint:cyclop
+	stuck := make(ItemList)
+
+	for idx, app := range c.Apps.Sonarr {
+		instance := idx + 1
+
+		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
+			continue
+		}
+
+		if app.Sonarr == nil {
+			c.Errorf("Getting Sonarr Queue (%d): Sonarr config is nil? This is probably a bug.", instance)
+			continue
+		}
+
+		start := time.Now()
+
+		queue, err := app.GetQueue(getItemsMax, 1)
+		if err != nil {
+			c.Errorf("Getting Sonarr Queue (%d): %v", instance, err)
+			continue
+		}
+
+		// repeatStomper is used to collapse duplicate download IDs.
+		repeatStomper := make(map[string]*sonarr.QueueRecord)
+		stuckapp := stuck[instance]
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			} else if repeatStomper[item.DownloadID] != nil {
+				continue
+			}
+
+			item.Quality = nil
+			item.Language = nil
+			repeatStomper[item.DownloadID] = item
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = app.Name
+		stuckapp.Elapsed = time.Since(start)
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Sonarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
+}
