+++ pkg/triggers/stuckitems/stuckitems.go
@@ -2,14 +2,12 @@ package stuckitems
 
 import (
 	"fmt"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/common"
 	"github.com/Notifiarr/notifiarr/pkg/website"
-	"golift.io/starr/sonarr"
 )
 
 /* This file contains the procedures to send stuck download queue items to notifiarr. */
@@ -173,196 +171,4 @@ func (c *cmd) getQueues() *QueuePayload {
 	wg.Wait()
 
 	return cue
-}
-
-func (c *cmd) getFinishedItemsLidarr() ItemList { //nolint:dupl,cyclop
-	stuck := make(ItemList)
-
-	for idx, app := range c.Apps.Lidarr {
-		instance := idx + 1
-
-		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
-			continue
-		}
-
-		if app.Lidarr == nil {
-			c.Errorf("Getting Lidarr Queue (%d): Lidarr config is nil? This is probably a bug.", instance)
-			continue
-		}
-
-		start := time.Now()
-
-		queue, err := app.GetQueue(getItemsMax, getItemsMax)
-		if err != nil {
-			c.Errorf("Getting Lidarr Queue (%d): %v", instance, err)
-			continue
-		}
-
-		stuckapp := stuck[instance]
-
-		for _, item := range queue.Records {
-			if s := strings.ToLower(item.Status); s != completed && s != warning &&
-				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
-				continue
-			}
-
-			item.Quality = nil
-			stuckapp.Queue = append(stuckapp.Queue, item)
-		}
-
-		stuckapp.Name = app.Name
-		stuckapp.Elapsed = time.Since(start)
-		stuck[instance] = stuckapp
-
-		c.Debugf("Checking Lidarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
-			instance, len(queue.Records), len(stuck[instance].Queue))
-	}
-
-	return stuck
-}
-
-func (c *cmd) getFinishedItemsRadarr() ItemList { //nolint:cyclop
-	stuck := make(ItemList)
-
-	for idx, app := range c.Apps.Radarr {
-		instance := idx + 1
-
-		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
-			continue
-		}
-
-		if app.Radarr == nil {
-			c.Errorf("Getting Radarr Queue (%d): Radarr config is nil? This is probably a bug.", instance)
-			continue
-		}
-
-		start := time.Now()
-
-		queue, err := app.GetQueue(getItemsMax, 1)
-		if err != nil {
-			c.Errorf("Getting Radarr Queue (%d): %v", instance, err)
-			continue
-		}
-
-		stuckapp := stuck[instance]
-
-		for _, item := range queue.Records {
-			if s := strings.ToLower(item.Status); s != completed && s != warning &&
-				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
-				continue
-			}
-
-			item.Quality = nil
-			item.CustomFormats = nil
-			item.Languages = nil
-			stuckapp.Queue = append(stuckapp.Queue, item)
-		}
-
-		stuckapp.Name = app.Name
-		stuckapp.Elapsed = time.Since(start)
-		stuck[instance] = stuckapp
-
-		c.Debugf("Checking Radarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
-			instance, len(queue.Records), len(stuck[instance].Queue))
-	}
-
-	return stuck
-}
-
-func (c *cmd) getFinishedItemsReadarr() ItemList { //nolint:dupl,cyclop
-	stuck := make(ItemList)
-
-	for idx, app := range c.Apps.Readarr {
-		instance := idx + 1
-
-		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
-			continue
-		}
-
-		if app.Readarr == nil {
-			c.Errorf("Getting Readarr Queue (%d): Readarr config is nil? This is probably a bug.", instance)
-			continue
-		}
-
-		start := time.Now()
-
-		queue, err := app.GetQueue(getItemsMax, getItemsMax)
-		if err != nil {
-			c.Errorf("Getting Readarr Queue (%d): %v", instance, err)
-			continue
-		}
-
-		stuckapp := stuck[instance]
-
-		for _, item := range queue.Records {
-			if s := strings.ToLower(item.Status); s != completed && s != warning &&
-				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
-				continue
-			}
-
-			item.Quality = nil
-			stuckapp.Queue = append(stuckapp.Queue, item)
-		}
-
-		stuckapp.Name = app.Name
-		stuckapp.Elapsed = time.Since(start)
-		stuck[instance] = stuckapp
-
-		c.Debugf("Checking Readarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
-			instance, len(queue.Records), len(stuck[instance].Queue))
-	}
-
-	return stuck
-}
-
-func (c *cmd) getFinishedItemsSonarr() ItemList { //nolint:cyclop
-	stuck := make(ItemList)
-
-	for idx, app := range c.Apps.Sonarr {
-		instance := idx + 1
-
-		if !app.StuckItem || app.URL == "" || app.APIKey == "" || app.Timeout.Duration < 0 {
-			continue
-		}
-
-		if app.Sonarr == nil {
-			c.Errorf("Getting Sonarr Queue (%d): Sonarr config is nil? This is probably a bug.", instance)
-			continue
-		}
-
-		start := time.Now()
-
-		queue, err := app.GetQueue(getItemsMax, 1)
-		if err != nil {
-			c.Errorf("Getting Sonarr Queue (%d): %v", instance, err)
-			continue
-		}
-
-		// repeatStomper is used to collapse duplicate download IDs.
-		repeatStomper := make(map[string]*sonarr.QueueRecord)
-		stuckapp := stuck[instance]
-
-		for _, item := range queue.Records {
-			if s := strings.ToLower(item.Status); s != completed && s != warning &&
-				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
-				continue
-			} else if repeatStomper[item.DownloadID] != nil {
-				continue
-			}
-
-			item.Quality = nil
-			item.Language = nil
-			repeatStomper[item.DownloadID] = item
-			stuckapp.Queue = append(stuckapp.Queue, item)
-		}
-
-		stuckapp.Name = app.Name
-		stuckapp.Elapsed = time.Since(start)
-		stuck[instance] = stuckapp
-
-		c.Debugf("Checking Sonarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
-			instance, len(queue.Records), len(stuck[instance].Queue))
-	}
-
-	return stuck
 }
